[Enhancement] adding unit testing documentation and hooks
