--- addon/adapter.js
@@ -1,10 +1,0 @@
-import Ember from 'ember';
-
-function notImplemented () {
-    throw new Error('not implemented');
-}
-
-export default Ember.Object.extend({
-    findLanguage:    notImplemented,
-    findTranslation: notImplemented
-});
